+++ bins/revme/src/cmd/bench/burntpix.rs
@@ -98,7 +98,7 @@ pub fn svg(filename: String, svg_data: &[u8]) -> Result<(), Box<dyn Error>> {
     let svg_dir = current_dir.join("burntpix").join("svgs");
     std::fs::create_dir_all(&svg_dir)?;
 
-    let file_path = svg_dir.join(format!("{}.svg", filename));
+    let file_path = svg_dir.join(format!("{filename}.svg"));
     let mut file = File::create(file_path)?;
     file.write_all(svg_data)?;
 
@@ -117,7 +117,7 @@ fn try_init_env_vars() -> Result<(u32, U256), Box<dyn Error>> {
 
 fn try_from_hex_to_u32(hex: &str) -> Result<u32, Box<dyn Error>> {
     let trimmed = hex.strip_prefix("0x").unwrap_or(hex);
-    u32::from_str_radix(trimmed, 16).map_err(|e| format!("Failed to parse hex: {}", e).into())
+    u32::from_str_radix(trimmed, 16).map_err(|e| format!("Failed to parse hex: {e}").into())
 }
 
 fn insert_account_info(cache_db: &mut CacheDB<EmptyDB>, addr: Address, code: &str) {