+++ examples/block_traces/src/main.rs
@@ -1,4 +1,6 @@
-//! Optimism-specific constants, types, and helpers.
+//! Example that show how to replay a block and trace the execution of each transaction.
+//!
+//! The EIP3155 trace of each transaction is saved into file `traces/{tx_number}.json`.
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
 use alloy_consensus::Transaction;
@@ -136,7 +138,7 @@ async fn main() -> anyhow::Result<()> {
             .build()
             .unwrap();
 
-        let file_name = format!("traces/{}.json", tx_number);
+        let file_name = format!("traces/{tx_number}.json");
         let write = OpenOptions::new()
             .write(true)
             .create(true)
@@ -151,7 +153,7 @@ async fn main() -> anyhow::Result<()> {
         let res: Result<_, _> = evm.inspect_one(tx, TracerEip3155::new(Box::new(writer)));
 
         if let Err(error) = res {
-            println!("Got error: {:?}", error);
+            println!("Got error: {error:?}");
         }
 
         // Flush the file writer